refactor(vscode): lift pure card helpers + error panel out of behavior.ts
Groundwork for the eventual <preview-card> migration in #767. Two new
typed modules with verbatim semantics:

- `cardData.ts` — pure data helpers: `sanitizeId`, `mimeFor`,
  `isAnimatedPreview`, `buildVariantLabel`, `shortDevice`,
  `isWearPreview`, `buildTooltip`, `parseBounds`. No DOM, no vscode
  handle — straight `PreviewInfo` → `string` / boolean / parsed
  bounds. Was previously eight `function foo(...) {}` blocks inside
  `setupPreviewBehavior` shielded by @ts-nocheck.
- `errorPanel.ts` — `buildErrorPanel(vscode, message, renderError,
  className)` returning the detached error overlay DOM. Caller hands
  in the vscode handle so the file stays untouched-by-closure.

`behavior.ts` shrinks from 3208 to 3051 lines (-157 LOC). All 320
tests pass; prettier clean. No behaviour change.

Refs #767.
